fix(cockpit): stop node:crypto leaking into the client bundle via the upload dropzone
upload-dropzone (client) → upload/batch → select/mappers → node:crypto.createHash
externalized the module for the browser and error-boundaried the whole cockpit on
load. Move the pure ext→source_type helper (sourceTypeForUri + its map) into the
crypto-free upload/policy; select/mappers re-exports it so server callers are
untouched, and upload/batch imports it from policy. Pre-existing (surfaced by the
first real-browser smoke since the UI went untested).
